split code in ass_format.c into 2 files (ass_parse.c) to separate the parsing functions (taken from libass) from newly added webVTT formatting, and for more readability. (#24) (#26)
