+++ config
@@ -399,6 +399,7 @@ VOD_SRCS="$VOD_SRCS                                           \
           $ngx_addon_dir/vod/mp4/mp4_parser_base.c            \
           $ngx_addon_dir/vod/mss/mss_packager.c               \
           $ngx_addon_dir/vod/subtitle/ass_format.c            \
+          $ngx_addon_dir/vod/subtitle/ass_parse.c             \
           $ngx_addon_dir/vod/subtitle/cap_format.c            \
           $ngx_addon_dir/vod/subtitle/subtitle_format.c       \
           $ngx_addon_dir/vod/subtitle/ttml_builder.c          \
